fix warnings

+++ WinCPURAM.cpp
@@ -23,37 +23,37 @@ WinCPURAM::WinCPURAM()
 unsigned int WinCPURAM::getTotalVirtualMemory()
 {
     GlobalMemoryStatusEx(&memInfo);
-    return memInfo.ullTotalPageFile;
+    return static_cast<unsigned int>(memInfo.ullTotalPageFile);
 }
 
 unsigned int WinCPURAM::getTotalVirtualMemoryUsed()
 {
     GlobalMemoryStatusEx(&memInfo);
-    return memInfo.ullTotalPageFile - memInfo.ullAvailPageFile;
+    return static_cast<unsigned int>(memInfo.ullTotalPageFile - memInfo.ullAvailPageFile);
 }
 
 unsigned int WinCPURAM::getTotalPhysicalMemory()
 {
     GlobalMemoryStatusEx(&memInfo);
-    return memInfo.ullTotalPhys;
+    return static_cast<unsigned int>(memInfo.ullTotalPhys);
 }
 
 unsigned int WinCPURAM::getTotalPhysicalMemoryUsed()
 {
     GlobalMemoryStatusEx(&memInfo);
-    return memInfo.ullTotalPhys - memInfo.ullAvailPhys;
+    return static_cast<unsigned int>(memInfo.ullTotalPhys - memInfo.ullAvailPhys);
 }
 
 unsigned int WinCPURAM::getVirtualMemory()
 {
     GetProcessMemoryInfo(GetCurrentProcess(), (PROCESS_MEMORY_COUNTERS*)&pmc, sizeof(pmc));
-    return  pmc.PrivateUsage/(1024*1024);
+    return  static_cast<unsigned int>(pmc.PrivateUsage/(1024*1024));
 }
 
 unsigned int WinCPURAM::getPhysicalMemory()
 {
     GetProcessMemoryInfo(GetCurrentProcess(), (PROCESS_MEMORY_COUNTERS*)&pmc, sizeof(pmc));
-    return  pmc.WorkingSetSize/(1024*1024);
+    return  static_cast<unsigned int>(pmc.WorkingSetSize/(1024*1024));
 }
 
 double WinCPURAM::getCPU()
@@ -68,8 +68,7 @@ double WinCPURAM::getCPU()
     GetProcessTimes(self, &ftime, &ftime, &fsys, &fuser);
     memcpy(&sys, &fsys, sizeof(FILETIME));
     memcpy(&user, &fuser, sizeof(FILETIME));
-    percent = (sys.QuadPart - lastSysCPU.QuadPart) +
-        (user.QuadPart - lastUserCPU.QuadPart);
+    percent = static_cast<double>((sys.QuadPart - lastSysCPU.QuadPart) + (user.QuadPart - lastUserCPU.QuadPart));
     percent /= (now.QuadPart - lastCPU.QuadPart);
     percent /= numProcessors;
     lastCPU = now;
